starting fresh, moved learning module service into its own folder under services
